+++ app/src/androidTest/java/sku/challenge/atmanatodoapp/ui/EditItemFragmentTest.kt
@@ -27,6 +27,7 @@ import sku.challenge.atmanatodoapp.repository.ItemRepository
 import sku.challenge.atmanatodoapp.test_utils.DataBindingIdlingResourceRule
 import sku.challenge.atmanatodoapp.test_utils.launchFragmentInHiltContainer
 import sku.challenge.atmanatodoapp.test_utils.mock
+import sku.challenge.atmanatodoapp.ui.edit_item.EditItemFragment
 import sku.challenge.atmanatodoapp.vo.Item
