+++ app/src/androidTest/java/sku/challenge/atmanatodoapp/ui/LocalFragmentTest.kt
@@ -27,6 +27,7 @@ import sku.challenge.atmanatodoapp.R
 import sku.challenge.atmanatodoapp.di.AppModule
 import sku.challenge.atmanatodoapp.repository.ItemRepository
 import sku.challenge.atmanatodoapp.test_utils.*
+import sku.challenge.atmanatodoapp.ui.local.LocalFragment
 import sku.challenge.atmanatodoapp.vo.Item
