+++ app/src/main/java/sku/challenge/atmanatodoapp/ui/remote/RemoteViewModel.kt
@@ -1,4 +1,4 @@
-package sku.challenge.atmanatodoapp.ui
+package sku.challenge.atmanatodoapp.ui.remote
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
