+++ app/src/test/java/sku/challenge/atmanatodoapp/ui/EditItemViewModelTest.kt
@@ -19,6 +19,7 @@ import org.mockito.Mockito.`when`
 import org.mockito.Mockito.verify
 import sku.challenge.atmanatodoapp.repository.ItemRepository
 import sku.challenge.atmanatodoapp.test_utils.mock
+import sku.challenge.atmanatodoapp.ui.edit_item.EditItemViewModel
 import sku.challenge.atmanatodoapp.vo.Item
