+++ app/src/test/java/sku/challenge/atmanatodoapp/ui/LocalViewModelTest.kt
@@ -16,6 +16,7 @@ import org.junit.Before
 import org.junit.Test
 import sku.challenge.atmanatodoapp.fake.FakeRepository
 import sku.challenge.atmanatodoapp.test_utils.DummyData
+import sku.challenge.atmanatodoapp.ui.local.LocalViewModel
 
 @ExperimentalCoroutinesApi
 class LocalViewModelTest {
